finalize

+++ codalab/migration.py
@@ -65,11 +65,16 @@ class Migration:
         assert StorageType.AZURE_BLOB_STORAGE.value == self.target_store['storage_type']
         self.target_store_type = StorageType.AZURE_BLOB_STORAGE
 
-    def get_bundle_uuids(self, worksheet_uuid):
-        bundle_uuids = self.bundle_manager._model.get_bundle_uuids(
-            {'name': None, 'worksheet_uuid': worksheet_uuid, 'user_id': self.root_user_id},
-            max_results=None,  # return all bundles in the worksheets
-        )
+    def get_bundle_uuids(self, worksheet_uuid, max_result=100):
+        if worksheet_uuid is None:
+            bundle_uuids = self.bundle_manager._model.get_all_bundle_uuids(
+                max_results=max_result
+            )
+        else:
+            bundle_uuids = self.bundle_manager._model.get_bundle_uuids(
+                {'name': None, 'worksheet_uuid': worksheet_uuid, 'user_id': self.root_user_id},
+                max_results=None,  # return all bundles in the worksheets
+            )
         return bundle_uuids
 
     def is_linked_bundle(self, bundle_uuid):
@@ -192,13 +197,12 @@ class Migration:
                 {'user_id': bundle_user_id, 'disk_used': new_disk_used}
             )
 
-
 if __name__ == '__main__':
     # Command line parser, parse the worksheet id
     parser = argparse.ArgumentParser(
         description='Manages your local CodaLab Worksheets service deployment'
     )
-    parser.add_argument('worksheet', type=str, help='The worksheet uuid that needs migration')
+    parser.add_argument('-w', '--worksheet', type=str, help='The worksheet uuid that needs migration')
     parser.add_argument('--target_store_name', type=str, help='The destination bundle store name')
     parser.add_argument(
         '-d',
@@ -216,7 +220,7 @@ if __name__ == '__main__':
     target_store_name = (
         "azure-store-default" if args.target_store_name is None else args.target_store_name
     )
-    if not spec_util.UUID_REGEX.match(worksheet_uuid):
+    if worksheet_uuid is not None and not spec_util.UUID_REGEX.match(worksheet_uuid):
         raise Exception("Input worksheet uuid has wrong format. ")
 
     # TODO: write output to log / log files
@@ -226,6 +230,8 @@ if __name__ == '__main__':
     bundle_uuids = migration.get_bundle_uuids(worksheet_uuid)
 
     for bundle_uuid in bundle_uuids:
+        logging.info(bundle_uuid)
+        print(bundle_uuid)
         bundle = migration.get_bundle(bundle_uuid)
         if bundle.bundle_type != 'dataset' or bundle.state != 'ready':
             # only migrate uploaded bundle, and the bundle state needs to be ready
@@ -248,6 +254,9 @@ if __name__ == '__main__':
 
         is_dir = bundle_info['type'] == 'directory'
         migration.upload_to_azure_blob(bundle_uuid, bundle_location, is_dir)
-        migration.modify_bundle_data(bundle, bundle_uuid, is_dir)
-        migration.sanity_check(bundle_uuid, bundle_location, bundle_info, is_dir)
-        migration.delete_original_bundle(bundle_uuid, bundle_location)
+        
+        if not args.dry_run:  # If dry_run, only upload to new bundle location
+            migration.modify_bundle_data(bundle, bundle_uuid, is_dir)
+            migration.sanity_check(bundle_uuid, bundle_location, bundle_info, is_dir)
+            if not args.keep:
+                migration.delete_original_bundle(bundle_uuid, bundle_location)
